Renamed 2 models (room_type and invoice_tax) to be camel case. Changed their references in their respective controllers.
