+++ controllers/invoicesController.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const Invoice = require('../models/invoice');
-const InvoiceTax = require('../models/invoice_tax');
+const InvoiceTax = require('../models/invoiceTax');
 const InvoicePayment = require('../models/invoicePayment');
 
 // all these routes point to /api/invoices as specified in server.js and controllers/index.js
